parent bc91a40c23cd62ec45ab1164f513ab2d20ac6309 author ppwadhwa <pwadhwa@quansight.com> 1568173352 -0500 committer ppwadhwa <pwadhwa@quansight.com> 1568948664 -0500
update .yml file to include sphinx (plus dependencies) and scikit-learn.  Also, created initial sphinx docs.

edited a few of the documents to be consistent on capitalization

remove line to add nbsite to path, as its already installed in the environment

make a few changes to base.py where special characters need to be ignored so sphinx is not confused