Refactor patitioning, and add partitioning CLI
This update creates the new DatasetPartitioner object,
which manages the repartitioning of butler-read data into
kartothek datasets, as well as writing tables of summary statistics.
This also adds a lsst_data_repartition CLI.
The only changes necessary to the dashboard code are renaming the datasets
in the kartothek reads,
as the dataset names now match the original butler dataset names.