persist to compute, and remove setindex as filter
Performance is much better when the active dataset is in direct RAM,
rather than persisted in RAM spread out on worker nodes.  For all practical use cases,
I believe this will be fine; i.e., we should not exceed head node RAM.
Also, this removes the setting the filter column as index, as this is time-consuming,
and (I think) unnecessary.